requirements.txt
fastapi
uvicorn
requests
pyjwt
python-dotenv
tenacity
